# agents/nightshift/gen/openapiv2/nightshift/v1/artifacts.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "nightshift/v1/artifacts.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/artifacts": {
      "get": {
        "summary": "ListArtifacts returns artifacts the caller has access to, with\noptional filters.",
        "operationId": "Artifacts_ListArtifacts",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListArtifactsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "typeFilter",
            "description": "Optional type filter. If set, only artifacts of this type are\nreturned.\n\n - ARTIFACT_TYPE_UNSPECIFIED: Proto3 default. Conforming implementations should not return this.\n - ARTIFACT_TYPE_OBJECT: Binary blob identified by a Storage Object.\n - ARTIFACT_TYPE_APP: Hosted static site deployed as a workload in the implementation's\nruntime.",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "ARTIFACT_TYPE_UNSPECIFIED",
              "ARTIFACT_TYPE_OBJECT",
              "ARTIFACT_TYPE_APP"
            ],
            "default": "ARTIFACT_TYPE_UNSPECIFIED"
          },
          {
            "name": "ownerId",
            "description": "Optional owner filter. If set, only artifacts owned by this user\nare returned. Permitted values depend on implementation auth:\ntypically a non-admin caller may only set this to their own id.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "runId",
            "description": "Optional run filter. If set, only artifacts produced during this\nrun are returned.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pageSize",
            "description": "Maximum number of artifacts to return. The server may cap the\nvalue.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageToken",
            "description": "Opaque pagination token from a previous ListArtifactsResponse.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "orderBy",
            "description": "Implementation-defined order specification. Conventional format is\n\"\u003cfield\u003e \u003casc|desc\u003e\", for example \"created_at desc\".",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "sessionId",
            "description": "Optional session filter. If set, only artifacts attached to this\nWorkers session are returned.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Artifacts"
        ]
      }
    },
...
    "/v1/artifacts/{artifactId}": {
      "get": {
        "summary": "GetArtifact reads an artifact by id. Permission is required — the\nRPC fails with NOT_FOUND when the caller is neither owner nor\ngrantee nor (for `public=true` artifacts) any authenticated user.",
        "operationId": "Artifacts_GetArtifact",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetArtifactResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "artifactId",
            "description": "Target artifact identifier.",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Artifacts"
        ]
      },
      "delete": {
        "summary": "DeleteArtifact removes an artifact, its permissions, its Storage\nObjects, and (for apps) its hosted workload. Owner only.",
        "operationId": "Artifacts_DeleteArtifact",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteArtifactResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "artifactId",
            "description": "Target artifact identifier.",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Artifacts"
        ]
      },
      "patch": {
        "summary": "UpdateArtifact modifies metadata, visibility, and/or bytes of an\nexisting artifact. Owner or editor role required; public/private\ntoggles are owner-only.",
        "operationId": "Artifacts_UpdateArtifact",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateArtifactResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "artifactId",
            "description": "Target artifact identifier.",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ArtifactsUpdateArtifactBody"
            }
          }
        ],
        "tags": [
          "Artifacts"
        ]
      }
    },
...
    "/v1/artifacts/{artifactId}:downloadUrl": {
      "get": {
        "summary": "GetArtifactDownloadURL returns a presigned URL the caller can GET\nto fetch the blob bytes of an object artifact. Any role (viewer,\neditor, owner) or public access suffices.",
        "operationId": "Artifacts_GetArtifactDownloadURL",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetArtifactDownloadURLResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "artifactId",
            "description": "Target artifact identifier. Must refer to an object artifact.",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Artifacts"
        ]
      }
    },
    "/v1/artifacts/{artifactId}:previewUrl": {
      "get": {
        "summary": "GetArtifactPreviewURL returns a presigned URL the caller can GET\nto fetch the companion HTML preview of an object artifact. Fails\nwith NOT_FOUND when the artifact has no preview.",
        "operationId": "Artifacts_GetArtifactPreviewURL",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetArtifactPreviewURLResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "artifactId",
            "description": "Target artifact identifier. Must refer to an object artifact with\n`has_preview` = true; otherwise the RPC fails with NOT_FOUND.",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Artifacts"
        ]
      }
    },
...
  },
  "definitions": {
...
    "ArtifactsUpdateArtifactBody": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "description": "Optional replacement display name."
        },
        "description": {
          "type": "string",
          "description": "Optional replacement description."
        },
        "public": {
          "type": "boolean",
          "description": "Optional new visibility. For app artifacts the implementation\nhandles the ingress toggle; for object artifacts this only changes\naccess enforcement, not the underlying Storage Object."
        },
        "htmlContent": {
          "type": "string",
          "description": "App-only: replacement HTML content. Ignored for objects."
        },
        "contentBytes": {
          "type": "string",
          "format": "byte",
          "description": "Object-only: replacement blob bytes."
        },
        "contentType": {
          "type": "string",
          "description": "Object-only: MIME type of `content_bytes`. Required whenever\n`content_bytes` is non-empty."
        },
        "previewHtml": {
          "type": "string",
          "format": "byte",
          "description": "Object-only: replacement companion preview HTML. Empty bytes\nexplicitly clears any prior preview."
        },
        "setPreviewHtml": {
          "type": "boolean",
          "description": "Whether `preview_html` is being set (true) or left alone (false).\nNeeded to distinguish \"clear the preview\" (set=true, bytes empty)\nfrom \"leave the preview\" (set=false)."
        }
      },
      "description": "UpdateArtifactRequest is the request message for UpdateArtifact.\n\nAny field left unset is preserved. Callers supply `name`,\n`description`, and/or `public` to change metadata; `content` (apps)\nor `content_bytes` + `content_type` (objects) to rewrite the bytes.\n`preview_html` follows `content_bytes` and may be set independently\nfor objects to ship a new preview without rewriting the binary."
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Artifact": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "Server-assigned identifier. Stable for the lifetime of the artifact."
        },
        "type": {
          "$ref": "#/definitions/v1ArtifactType",
          "description": "Whether this artifact is an object (blob) or an app (hosted static\nsite)."
        },
        "name": {
          "type": "string",
          "description": "Human-readable display name. For object artifacts this is typically\nthe filename."
        },
        "description": {
          "type": "string",
          "description": "Optional short description supplied by the creator."
        },
        "ownerId": {
          "type": "string",
          "description": "Identifier of the user who created the artifact. Treated opaquely;\nimplementations bind it to their auth identity model."
        },
        "runId": {
          "type": "string",
          "description": "Identifier of the run during which this artifact was produced, or\nempty if the artifact was created outside a run (for example via a\ndirect user upload)."
        },
        "public": {
          "type": "boolean",
          "description": "When false (default) access is gated by ownership + permission\ngrants. When true, access is unauthenticated."
        },
        "objectBucket": {
          "type": "string",
          "description": "Object-only: the Storage Object holding the blob bytes. Empty for\napp artifacts."
        },
        "objectKey": {
          "type": "string"
        },
        "contentType": {
          "type": "string",
          "description": "Object-only: MIME type of the blob bytes. Empty for app artifacts."
        },
        "sizeBytes": {
          "type": "string",
          "format": "int64",
          "description": "Object-only: size of the blob in bytes, as reported by Storage\nafter finalization. Zero for app artifacts."
        },
        "hasPreview": {
          "type": "boolean",
          "description": "Object-only: whether a companion preview HTML was uploaded to the\nimplementation's preview key convention for this artifact. When\ntrue, GetArtifactPreviewURL will return a URL."
        },
        "appUrl": {
          "type": "string",
          "description": "App-only: URL at which the hosted workload serves content. For\nprivate apps this is the implementation's permission-gated proxy\nURL; for public apps this is the public ingress URL. Empty for\nobject artifacts."
        },
        "appStatus": {
          "$ref": "#/definitions/v1DeploymentState",
          "description": "App-only: deployment lifecycle state. DEPLOYMENT_STATE_UNSPECIFIED\nfor object artifacts."
        },
        "appStatusDetail": {
          "type": "string",
          "description": "App-only: optional free-form reason string accompanying FAILED\nstate."
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "description": "Time the artifact record was first created."
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "description": "Time the artifact record was most recently updated."
        },
        "sessionId": {
          "type": "string",
          "description": "Workers session this artifact is attached to, or empty. Typically\nset on user-uploaded artifacts and left empty on worker-produced\nartifacts (which are scoped via run_id)."
        }
      },
      "description": "Artifact is the metadata record for an agent-produced or user-uploaded\noutput. The blob bytes (for objects) and the hosted workload (for\napps) live elsewhere; this message describes what, who, and where."
    },
...
    "v1ArtifactType": {
      "type": "string",
      "enum": [
        "ARTIFACT_TYPE_UNSPECIFIED",
        "ARTIFACT_TYPE_OBJECT",
        "ARTIFACT_TYPE_APP"
      ],
      "default": "ARTIFACT_TYPE_UNSPECIFIED",
      "description": "ArtifactType distinguishes object artifacts (blob content served via\ndownload URL) from app artifacts (static sites served as a hosted\nworkload).\n\n - ARTIFACT_TYPE_UNSPECIFIED: Proto3 default. Conforming implementations should not return this.\n - ARTIFACT_TYPE_OBJECT: Binary blob identified by a Storage Object.\n - ARTIFACT_TYPE_APP: Hosted static site deployed as a workload in the implementation's\nruntime."
    },
...
    "v1DeleteArtifactResponse": {
      "type": "object",
      "description": "DeleteArtifactResponse is the response message for DeleteArtifact. It\nis intentionally empty."
    },
    "v1DeploymentState": {
      "type": "string",
      "enum": [
        "DEPLOYMENT_STATE_UNSPECIFIED",
        "DEPLOYMENT_STATE_PENDING",
        "DEPLOYMENT_STATE_READY",
        "DEPLOYMENT_STATE_FAILED"
      ],
      "default": "DEPLOYMENT_STATE_UNSPECIFIED",
      "description": "DeploymentState is the lifecycle of an app artifact's hosted workload.\n\n - DEPLOYMENT_STATE_UNSPECIFIED: Proto3 default. Conforming implementations should not return this\nfor app artifacts once the initial deploy has started.\n - DEPLOYMENT_STATE_PENDING: The deploy has been accepted and the workload is being created.\n - DEPLOYMENT_STATE_READY: The workload is reachable at `app_url`.\n - DEPLOYMENT_STATE_FAILED: The most recent deploy or update failed. `app_status_detail` carries\nan implementation-specific reason."
    },
    "v1GetArtifactDownloadURLResponse": {
      "type": "object",
      "properties": {
        "downloadUrl": {
          "type": "string",
          "description": "Presigned URL the caller can GET to fetch the blob bytes."
        },
        "expiresAt": {
          "type": "string",
          "format": "date-time",
          "description": "When the URL becomes invalid."
        }
      },
      "description": "GetArtifactDownloadURLResponse is the response message for\nGetArtifactDownloadURL."
    },
    "v1GetArtifactPreviewURLResponse": {
      "type": "object",
      "properties": {
        "previewUrl": {
          "type": "string",
          "description": "Presigned URL the caller can GET to fetch the companion HTML\npreview."
        },
        "expiresAt": {
          "type": "string",
          "format": "date-time",
          "description": "When the URL becomes invalid."
        }
      },
      "description": "GetArtifactPreviewURLResponse is the response message for\nGetArtifactPreviewURL."
    },
    "v1GetArtifactResponse": {
      "type": "object",
      "properties": {
        "artifact": {
          "$ref": "#/definitions/v1Artifact",
          "description": "The requested artifact."
        }
      },
      "description": "GetArtifactResponse is the response message for GetArtifact."
    },
...
    "v1ListArtifactsResponse": {
      "type": "object",
      "properties": {
        "artifacts": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Artifact"
          },
          "description": "Matching artifacts on this page."
        },
        "nextPageToken": {
          "type": "string",
          "description": "Opaque pagination token. Empty when there are no further results."
        }
      },
      "description": "ListArtifactsResponse is the response message for ListArtifacts."
    },
...
    "v1UpdateArtifactResponse": {
      "type": "object",
      "properties": {
        "artifact": {
          "$ref": "#/definitions/v1Artifact",
          "description": "The artifact after the update."
        }
      },
      "description": "UpdateArtifactResponse is the response message for UpdateArtifact."
    }
  }
}